上传模块参数未配置时允许删除(只删除本地记录)

+++ extend/oss/aliyun/Upload.php
@@ -78,7 +78,7 @@ class Upload extends UploadBase
                     throw new OssException('解析数据失败');
                 }
             } catch (OssException $e) {
-                return $this->setError($e->getErrorMessage());
+                return $this->setError($e->getMessage());
             }
 
             $random = array_rand($result, 1);
@@ -526,11 +526,15 @@ class Upload extends UploadBase
         $endPoint = Config::get('aliyun_endpoint.value', 'upload');
         $bucket = Config::get('aliyun_bucket.value', 'upload');
 
+        if (empty($accessKeyId) || empty($accessKeySecret) || empty($endPoint) || empty($bucket)) {
+            return true;
+        }
+
         try {
             $ossClient = new OssClient($accessKeyId, $accessKeySecret, $endPoint);
             $ossClient->deleteObjects($bucket, $this->delFileList);
         } catch (OssException $e) {
-            return $this->setError($e->getErrorMessage());
+            return $this->setError($e->getMessage());
         }
 
         return true;
